auth-service/requirements.txt
fastapi==0.104.3
uvicorn[standard]==0.23.1
pydantic==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==1.3.1
